+++ cypress/cypress/component/FormGroupLayout/FormGroupLayout.cy.tsx
@@ -77,6 +77,54 @@ it("testing existing * on nested object", () => {
   cy.get(`label[for=${fakePerson.city.address.street}]`).should("have.text", `${fakePerson.city.address.street} *`);
 });
 
+it("testing existing * on arrays", () => {
+  const fakePerson = {
+    city: {
+      address: {
+        streetsAsString: [faker.random.alpha(10), faker.random.alpha(10), faker.random.alpha(10)],
+        streetsAsObject: [
+          { name: faker.random.alpha(10), date: new Date() },
+          { name: faker.random.alpha(10), date: new Date() },
+          { name: faker.random.alpha(10), date: new Date() },
+        ],
+      },
+    },
+  };
+
+  const schema = yup.object().shape({
+    city: yup.object().shape({
+      address: yup.object().shape({
+        streetsAsString: yup.array().of(yup.string().required()),
+        streetsAsObject: yup.array().of(
+          yup.object().shape({
+            name: yup.string().required(),
+            date: yup.date(),
+          }),
+        ),
+      }),
+    }),
+  });
+
+  mount(
+    <Form<typeof fakePerson>
+      onSubmit={() => {
+        // Nothing to do
+      }}
+      defaultValues={fakePerson}
+      resolver={yupResolver(schema)}
+      requiredFields={["city.address.streetsAsString", "city.address.streetsAsObject.*.name", "city.address.streetsAsObject.*.date"]}
+    >
+      <Input<typeof fakePerson> name="city.address.streetsAsObject.0.name" label="Street as object" />
+      <Input<typeof fakePerson> name="city.address.streetsAsString.0" label="Street as string" />
+      <DatePickerInput<typeof fakePerson> name="city.address.streetsAsObject.0.date" label="Date as object" />
+    </Form>,
+  );
+
+  cy.get(`label[for="city.address.streetsAsObject.0.name"`).should("have.text", "Street as object *");
+  cy.get(`label[for="city.address.streetsAsString.0"`).should("have.text", "Street as string *");
+  cy.get(`label[for="city.address.streetsAsObject.0.date"`).should("have.text", "Date as object *");
+});
+
 const ValidationForm = (props: { hideValidationMessage?: boolean; hideValidationMessages?: boolean }) => {
   const { hideValidationMessage, hideValidationMessages } = props;
   const schema = yup.object().shape({